chore(test): expand eslint no-direct-db-in-tests rule to api-test-helpers
Fix remaining DB-direct violations in credits.ts (3 transaction wrappers)
and schedules.ts (resolveAgentIdFromCompose + initServices calls). Expand
the eslint rule to cover api-test-helpers/**/*.ts and add db/schema/* import
detection to prevent future regressions.

After this change, all files under api-test-helpers/ have zero
globalThis.services.db, db/schema/*, and initServices() references.

Closes #9381
