+++ turbo/apps/web/src/__tests__/api-test-helpers/credits.ts
@@ -1,10 +1,3 @@
-import { initServices } from "../../lib/init-services";
-import { grantOrgCredits } from "../../lib/zero/org/org-service";
-import {
-  deductFromExpiresRecords,
-  expireCredits,
-} from "../../lib/zero/credit/credit-expires-service";
-
 // ---------------------------------------------------------------------------
 // Re-exports: DB-direct seeders.
 //
@@ -26,6 +19,9 @@ export {
   seedCreditUsageRecord,
   seedInsightsDaily,
   createCompletedRun,
+  grantCreditsToOrg,
+  testDeductFromExpiresRecords,
+  testExpireCredits,
 } from "../db-test-seeders/credits";
 
 // ---------------------------------------------------------------------------
@@ -42,49 +38,3 @@ export {
   findUsageDaily,
   findInsightsDaily,
 } from "../db-test-assertions/credits";
-
-// ---------------------------------------------------------------------------
-// Service-layer wrappers.
-//
-// These call production service functions (not raw DB) and are valid
-// API-based helpers.
-// ---------------------------------------------------------------------------
-
-/**
- * Grant credits to an org atomically. Wraps grantOrgCredits in a transaction.
- */
-export async function grantCreditsToOrg(
-  orgId: string,
-  amount: number,
-): Promise<void> {
-  initServices();
-  await globalThis.services.db.transaction(async (tx) => {
-    await grantOrgCredits(tx, orgId, amount);
-  });
-}
-
-/**
- * Deduct from expires records within a transaction (test helper).
- */
-export async function testDeductFromExpiresRecords(
-  orgId: string,
-  amount: number,
-): Promise<void> {
-  initServices();
-  await globalThis.services.db.transaction(async (tx) => {
-    await deductFromExpiresRecords(tx, orgId, amount);
-  });
-}
-
-/**
- * Expire credits within a transaction (test helper).
- * Returns the total expired amount.
- */
-export async function testExpireCredits(orgId: string): Promise<number> {
-  initServices();
-  let result = 0;
-  await globalThis.services.db.transaction(async (tx) => {
-    result = await expireCredits(tx, orgId);
-  });
-  return result;
-}
