+++ turbo/apps/web/src/__tests__/db-test-seeders/schedules.ts
@@ -1,6 +1,8 @@
 import { and, eq } from "drizzle-orm";
 import { initServices } from "../../lib/init-services";
 import { zeroAgentSchedules } from "../../db/schema/zero-agent-schedule";
+import { agentComposes } from "../../db/schema/agent-compose";
+import { zeroAgents } from "../../db/schema/zero-agent";
 
 // ============================================================================
 // Schedule Seeders
@@ -75,4 +77,37 @@ export async function setScheduleConsecutiveFailures(
         eq(zeroAgentSchedules.name, name),
       ),
     );
+}
+
+/**
+ * Resolve composeId to agentId for test helpers.
+ * Looks up the compose to get org/name, then finds the corresponding zero agent.
+ *
+ * @why-db-direct Resolves compose → zero_agent mapping; no API endpoint
+ * provides composeId→agentId resolution
+ */
+export async function resolveAgentIdFromCompose(
+  composeId: string,
+): Promise<string> {
+  initServices();
+  const [compose] = await globalThis.services.db
+    .select({ orgId: agentComposes.orgId, name: agentComposes.name })
+    .from(agentComposes)
+    .where(eq(agentComposes.id, composeId))
+    .limit(1);
+  if (!compose) throw new Error(`Compose ${composeId} not found`);
+
+  const [agent] = await globalThis.services.db
+    .select({ id: zeroAgents.id })
+    .from(zeroAgents)
+    .where(
+      and(
+        eq(zeroAgents.orgId, compose.orgId),
+        eq(zeroAgents.name, compose.name),
+      ),
+    )
+    .limit(1);
+  if (!agent) throw new Error(`Zero agent not found for compose ${composeId}`);
+
+  return agent.id;
 }
